+++ apps/api/src/console.ts
@@ -2,12 +2,14 @@ import "reflect-metadata";
 import "@akashnetwork/env-loader";
 import "./open-telemetry";
 
+import { LoggerService } from "@akashnetwork/logging";
 import { context, trace } from "@opentelemetry/api";
 import { Command } from "commander";
 import { container } from "tsyringe";
 
 import { WalletController } from "@src/billing/controllers/wallet/wallet.controller";
-import { LoggerService } from "@src/core";
+import { chainDb } from "@src/db/dbConnection";
+import { TopUpDeploymentsController } from "@src/deployment/controllers/deployment/deployment.controller";
 
 const program = new Command();
 
@@ -17,19 +19,36 @@ const tracer = trace.getTracer("API Console");
 program
   .command("refill-wallets")
   .description("Refill draining wallets")
-  .action(async () => {
-    await context.with(trace.setSpan(context.active(), tracer.startSpan("refill-wallets")), async () => {
-      const logger = new LoggerService({ context: "CLI" });
-      logger.info("Refilling wallets");
-      // eslint-disable-next-line @typescript-eslint/no-var-requires
-      const { migratePG, closeConnections } = await require("./core/providers/postgres.provider");
-      await migratePG();
-
+  .action(async (options, command) => {
+    await executeCliHandler(command.name(), async () => {
       await container.resolve(WalletController).refillWallets();
-
-      await closeConnections();
-      logger.info("Finished refilling wallets");
     });
   });
+
+program
+  .command("top-up-deployments")
+  .description("Refill deployments with auto top up enabled")
+  .action(async (options, command) => {
+    await executeCliHandler(command.name(), async () => {
+      await container.resolve(TopUpDeploymentsController).topUpDeployments();
+    });
+  });
+
+async function executeCliHandler(name: string, handler: () => Promise<void>) {
+  await context.with(trace.setSpan(context.active(), tracer.startSpan(name)), async () => {
+    const logger = new LoggerService({ context: "CLI" });
+    logger.info({ event: "COMMAND_START", name });
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    const { migratePG, closeConnections } = await require("./core/providers/postgres.provider");
+    await migratePG();
+    await chainDb.authenticate();
+
+    await handler();
+
+    await closeConnections();
+    await chainDb.close();
+    logger.info({ event: "COMMAND_END", name });
+  });
+}
 
 program.parse();