+++ apps/api/src/core/providers/index.ts
@@ -1,2 +1,3 @@
 export * from "./postgres.provider";
 export * from "./config.provider";
+import "./http-sdk.provider";