+++ apps/deploy-web/src/components/layout/ModeToggle.tsx
@@ -2,10 +2,9 @@
 import * as React from "react";
 import { useEffect, useState } from "react";
 import { Button, DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@akashnetwork/ui/components";
+import { cn } from "@akashnetwork/ui/utils";
 import { HalfMoon, SunLight } from "iconoir-react";
 import { useTheme } from "next-themes";
-
-import { cn } from "@src/utils/styleUtils";
 
 export function ModeToggle() {
   const { setTheme, theme } = useTheme();