+++ apps/deploy-web/src/components/new-deployment/ManifestEdit.tsx
@@ -1,14 +1,16 @@
 "use client";
 import { Dispatch, SetStateAction, useEffect, useRef, useState } from "react";
 import { certificateManager } from "@akashnetwork/akashjs/build/certificates/certificate-manager";
-import { Alert, Button, CustomTooltip, Input, Spinner } from "@akashnetwork/ui/components";
+import { Alert, Button, CustomTooltip, FileButton, Input, Snackbar, Spinner } from "@akashnetwork/ui/components";
+import { cn } from "@akashnetwork/ui/utils";
 import { EncodeObject } from "@cosmjs/proto-signing";
 import { useTheme as useMuiTheme } from "@mui/material/styles";
 import useMediaQuery from "@mui/material/useMediaQuery";
-import { ArrowRight, InfoCircle } from "iconoir-react";
+import { ArrowRight, InfoCircle, Upload } from "iconoir-react";
 import { useAtom } from "jotai";
 import { useRouter, useSearchParams } from "next/navigation";
 import { event } from "nextjs-google-analytics";
+import { useSnackbar } from "notistack";
 
 import { browserEnvConfig } from "@src/config/browser-env.config";
 import { useCertificate } from "@src/context/CertificateProvider";
@@ -28,7 +30,6 @@ import { deploymentData } from "@src/utils/deploymentData";
 import { saveDeploymentManifestAndName } from "@src/utils/deploymentLocalDataUtils";
 import { validateDeploymentData } from "@src/utils/deploymentUtils";
 import { importSimpleSdl } from "@src/utils/sdl/sdlImport";
-import { cn } from "@src/utils/styleUtils";
 import { Timer } from "@src/utils/timer";
 import { TransactionMessageData } from "@src/utils/TransactionMessageData";
 import { domainName, handleDocClick, UrlService } from "@src/utils/urlUtils";
@@ -47,15 +48,23 @@ type Props = {
   selectedTemplate: TemplateCreation | null;
   editedManifest: string | null;
   setEditedManifest: Dispatch<SetStateAction<string>>;
+  isGitProviderTemplate?: boolean;
 };
 
-export const ManifestEdit: React.FunctionComponent<Props> = ({ editedManifest, setEditedManifest, onTemplateSelected, selectedTemplate }) => {
+export const ManifestEdit: React.FunctionComponent<Props> = ({
+  editedManifest,
+  setEditedManifest,
+  onTemplateSelected,
+  selectedTemplate,
+  isGitProviderTemplate
+}) => {
   const [parsingError, setParsingError] = useState<string | null>(null);
   const [deploymentName, setDeploymentName] = useState("");
   const [isCreatingDeployment, setIsCreatingDeployment] = useState(false);
   const [isDepositingDeployment, setIsDepositingDeployment] = useState(false);
   const [isCheckingPrerequisites, setIsCheckingPrerequisites] = useState(false);
   const [selectedSdlEditMode, setSelectedSdlEditMode] = useAtom(sdlStore.selectedSdlEditMode);
+  const [isRepoInputValid, setIsRepoInputValid] = useState(false);
   const [sdlDenom, setSdlDenom] = useState("uakt");
   const { settings } = useSettings();
   const { address, signAndBroadcastTx, isManaged } = useWallet();
@@ -71,10 +80,10 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({ editedManifest, s
   const templateId = searchParams.get("templateId");
   const { data: depositParams } = useDepositParams();
   const defaultDeposit = depositParams || browserEnvConfig.NEXT_PUBLIC_DEFAULT_INITIAL_DEPOSIT;
-  const fileUploadRef = useRef<HTMLInputElement>(null);
   const wallet = useWallet();
   const managedDenom = useManagedWalletDenom();
   const { createDeploymentConfirm } = useManagedDeploymentConfirm();
+  const { enqueueSnackbar } = useSnackbar();
 
   useWhen(
     wallet.isManaged && sdlDenom === "uakt" && editedManifest,
@@ -84,40 +93,13 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({ editedManifest, s
     },
     [editedManifest, wallet.isManaged, sdlDenom]
   );
-
   useWhen(hasComponent("ssh"), () => {
     setSelectedSdlEditMode("builder");
   });
 
-  const propagateUploadedSdl = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const selectedFiles = event.target.files ?? [];
-    const hasFileSelected = selectedFiles.length > 0;
-    if (!hasFileSelected) return;
-    const fileUploaded = selectedFiles[0];
-
-    const reader = new FileReader();
-
-    reader.onload = event => {
-      onTemplateSelected({
-        title: "From file",
-        code: "from-file",
-        category: "General",
-        description: "Custom uploaded file",
-        content: event.target?.result as string
-      });
-      setEditedManifest(event.target?.result as string);
-      setSelectedSdlEditMode("yaml");
-    };
-
-    reader.readAsText(fileUploaded);
-  };
-
-  const triggerFileInput = () => {
-    if (fileUploadRef.current) {
-      fileUploadRef.current.value = "";
-      fileUploadRef.current.click();
-    }
-  };
+  useWhen(isGitProviderTemplate, () => {
+    setSelectedSdlEditMode("builder");
+  }, [isGitProviderTemplate]);
 
   useEffect(() => {
     if (selectedTemplate?.name) {
@@ -141,6 +123,26 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({ editedManifest, s
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [editedManifest]);
+
+  const onFileSelect = (file: File | null) => {
+    if (!file) return;
+
+    const reader = new FileReader();
+
+    reader.onload = event => {
+      onTemplateSelected({
+        title: "From file",
+        code: "from-file",
+        category: "General",
+        description: "Custom uploaded file",
+        content: event.target?.result as string
+      });
+      setEditedManifest(event.target?.result as string);
+      setSelectedSdlEditMode("yaml");
+    };
+
+    reader.readAsText(file);
+  };
 
   async function handleTextChange(value) {
     setEditedManifest(value);
@@ -176,6 +178,13 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({ editedManifest, s
   }
 
   const handleCreateDeployment = async () => {
+    if (isGitProviderTemplate && !isRepoInputValid) {
+      enqueueSnackbar(<Snackbar title={"Please Fill All Required Fields"} subTitle="You need fill repo url and branch to deploy" iconVariant="error" />, {
+        variant: "error"
+      });
+      return;
+    }
+
     if (selectedSdlEditMode === "builder") {
       const valid = await sdlBuilderRef.current?.validate();
       if (!valid) return;
@@ -350,44 +359,46 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({ editedManifest, s
         </div>
       </div>
 
-      <div className="mb-2 flex gap-2">
-        {hasComponent("yml-editor") && (
-          <div className="flex items-center">
-            <Button
-              variant={selectedSdlEditMode === "builder" ? "default" : "outline"}
-              onClick={() => changeMode("builder")}
-              size="sm"
-              className={cn("flex-grow sm:flex-grow-0", { "rounded-e-none": hasComponent("yml-editor") })}
-              disabled={!!parsingError && selectedSdlEditMode === "yaml"}
-            >
-              Builder
-            </Button>
-            <Button
-              variant={selectedSdlEditMode === "yaml" ? "default" : "outline"}
-              color={selectedSdlEditMode === "yaml" ? "secondary" : "primary"}
-              onClick={() => changeMode("yaml")}
-              size="sm"
-              className="flex-grow rounded-s-none sm:flex-grow-0"
-            >
-              YAML
-            </Button>
-          </div>
-        )}
-        {hasComponent("yml-uploader") && !templateId && (
-          <>
-            <input type="file" ref={fileUploadRef} onChange={propagateUploadedSdl} style={{ display: "none" }} accept=".yml,.yaml,.txt" />
-            <Button
-              variant="outline"
-              color="secondary"
-              onClick={() => triggerFileInput()}
-              size="sm"
-              className="flex-grow hover:bg-primary hover:text-white sm:flex-grow-0"
-            >
-              Upload SDL
-            </Button>
-          </>
-        )}
-      </div>
+      {!isGitProviderTemplate && (
+        <div className="mb-2 flex gap-2">
+          {hasComponent("yml-editor") && (
+            <div className="flex items-center">
+              <Button
+                variant={selectedSdlEditMode === "builder" ? "default" : "outline"}
+                onClick={() => changeMode("builder")}
+                size="sm"
+                className={cn("flex-grow sm:flex-grow-0", { "rounded-e-none": hasComponent("yml-editor") })}
+                disabled={!!parsingError && selectedSdlEditMode === "yaml"}
+              >
+                Builder
+              </Button>
+              <Button
+                variant={selectedSdlEditMode === "yaml" ? "default" : "outline"}
+                color={selectedSdlEditMode === "yaml" ? "secondary" : "primary"}
+                onClick={() => changeMode("yaml")}
+                size="sm"
+                className="flex-grow rounded-s-none sm:flex-grow-0"
+              >
+                YAML
+              </Button>
+            </div>
+          )}
+          {hasComponent("yml-uploader") && !templateId && (
+            <>
+              <FileButton
+                onFileSelect={onFileSelect}
+                accept=".yml,.yaml,.txt"
+                size="sm"
+                variant="outline"
+                className="flex-grow hover:bg-primary hover:text-white sm:flex-grow-0"
+              >
+                <Upload className="text-xs" />
+                <span className="text-xs">Upload your SDL</span>
+              </FileButton>
+            </>
+          )}
+        </div>
+      )}
 
       {parsingError && <Alert variant="warning">{parsingError}</Alert>}
 
@@ -397,7 +408,15 @@ export const ManifestEdit: React.FunctionComponent<Props> = ({ editedManifest, s
         </ViewPanel>
       )}
       {(hasComponent("ssh") || selectedSdlEditMode === "builder") && (
-        <SdlBuilder sdlString={editedManifest} ref={sdlBuilderRef} setEditedManifest={setEditedManifest} />
+        <SdlBuilder
+          sdlString={editedManifest}
+          ref={sdlBuilderRef}
+          isGitProviderTemplate={isGitProviderTemplate}
+          setEditedManifest={setEditedManifest}
+          setDeploymentName={setDeploymentName}
+          deploymentName={deploymentName}
+          setIsRepoInputValid={setIsRepoInputValid}
+        />
       )}
 
       {isDepositingDeployment && (