+++ apps/deploy-web/src/components/providers/Uptime.tsx
@@ -1,7 +1,6 @@
 "use client";
 import { useIntl } from "react-intl";
-
-import { cn } from "@src/utils/styleUtils";
+import { cn } from "@akashnetwork/ui/utils";
 
 type Props = {
   value: number;