+++ apps/deploy-web/src/components/templates/TemplateBox.tsx
@@ -1,11 +1,11 @@
 "use client";
 import { Avatar, AvatarFallback, AvatarImage, cardClasses, CardContent, CardHeader } from "@akashnetwork/ui/components";
+import { cn } from "@akashnetwork/ui/utils";
 import { MediaImage } from "iconoir-react";
 import Link from "next/link";
 
 import { ApiTemplate } from "@src/types";
 import { getShortText } from "@src/utils/stringUtils";
-import { cn } from "@src/utils/styleUtils";
 import { UrlService } from "@src/utils/urlUtils";
 
 type Props = {
@@ -23,17 +23,17 @@ export const TemplateBox: React.FunctionComponent<Props> = ({ template, linkHref
       <CardHeader>
         <div className="flex items-center">
           <Avatar className="h-10 w-10">
-            <AvatarImage src={template.logoUrl} alt={template.name} />
+            <AvatarImage src={template.logoUrl} alt={template.name} className="object-contain" />
             <AvatarFallback>
               <MediaImage />
             </AvatarFallback>
           </Avatar>
 
-          <div className="ml-4 break-all font-bold tracking-tight">{template.name}</div>
+          <div className="ml-4 truncate text-nowrap font-bold tracking-tight">{template.name}</div>
         </div>
       </CardHeader>
       <CardContent className="pb-4 pt-0">
-        <p className="text-xs text-muted-foreground">{getShortText(template.summary, 128)}</p>
+        <p className="text-xs text-muted-foreground">{getShortText(template.summary, 80)}</p>
       </CardContent>
     </Link>
   );