+++ apps/deploy-web/src/lib/XTerm/XTerm.tsx
@@ -3,12 +3,12 @@ import "xterm/css/xterm.css";
 
 import { Ref, useEffect, useRef } from "react";
 import React from "react";
+import { cn } from "@akashnetwork/ui/utils";
 import { useTheme } from "next-themes";
 import { ITerminalAddon, ITerminalOptions, Terminal } from "xterm";
 import { FitAddon } from "xterm-addon-fit";
 
 import { copyTextToClipboard } from "@src/utils/copyClipboard";
-import { cn } from "@src/utils/styleUtils";
 
 export interface IProps {
   /**